+++ src/components/Layout/index.jsx
@@ -1,30 +1,69 @@
 import React from 'react';
-// import Container from '@material-ui/core/Container';
-// import CssBaseline from '@material-ui/core/CssBaseline';
-// import Grid from '@material-ui/core/Grid';
-import { GoogleNewsfeed } from '../../features/google-newsfeed';
-import { TwitterFeed } from '../../features/twitter-feed';
-import { RedditFeed } from '../../features/reddit-feed';
-import Grid from '@material-ui/core/Grid';
-import { View } from './view';
+import Drawer from '@material-ui/core/Drawer';
+import AppBar from '@material-ui/core/AppBar';
+import CssBaseline from '@material-ui/core/CssBaseline';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
+import { VerticalTabs } from 'components/VerticalTabs';
+import { makeStyles } from '@material-ui/core/styles';
 
-export function Layout() {
-  const renderNewsfeed = () => {
-    return (
-      <div style={{ width: 450 }}>
-        <Grid
-          container
-          direction="column"
-          justify="space-between"
-          alignItems="center"
-        >
-          <GoogleNewsfeed />
-          <TwitterFeed />
-          <RedditFeed />
-        </Grid>
-      </div>
-    );
-  };
+const drawerWidth = 450;
 
-  return <View renderToolbar={renderNewsfeed} />;
+const useStyles = makeStyles((theme) => ({
+  root: {
+    display: 'flex',
+  },
+  appBar: {
+    zIndex: theme.zIndex.drawer + 1,
+  },
+  drawer: {
+    width: drawerWidth,
+    flexShrink: 0,
+  },
+  drawerPaper: {
+    width: drawerWidth,
+  },
+  drawerContainer: {
+    overflow: 'auto',
+  },
+  main: {
+    flexGrow: 1,
+  },
+  content: {},
+}));
+
+export function Layout({ title, children, renderToolbar }) {
+  const classes = useStyles();
+
+  return (
+    <div className={classes.root}>
+      <CssBaseline />
+      <AppBar position="fixed" className={classes.appBar}>
+        <Toolbar>
+          <Typography variant="h6" noWrap>
+            {title}
+          </Typography>
+        </Toolbar>
+      </AppBar>
+      <main className={classes.main}>
+        <Toolbar />
+        <div className={classes.content}>
+          { children }
+        </div>
+      </main>
+      <Drawer
+        className={classes.drawer}
+        variant="permanent"
+        classes={{
+          paper: classes.drawerPaper,
+        }}
+        anchor="right"
+      >
+        <Toolbar />
+        <div className={classes.drawerContainer}>
+          {renderToolbar && renderToolbar()}
+        </div>
+      </Drawer>
+    </div>
+  );
 }
