--- src/components/Layout/styles.js
@@ -1,25 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-const drawerWidth = 240;
-export const useStyles = makeStyles((theme) => ({
-  root: {
-    display: 'flex',
-  },
-  appBar: {
-    zIndex: theme.zIndex.drawer + 1,
-  },
-  drawer: {
-    width: drawerWidth,
-    flexShrink: 0,
-  },
-  drawerPaper: {
-    width: drawerWidth,
-  },
-  drawerContainer: {
-    overflow: 'auto',
-  },
-  main: {
-    flexGrow: 1,
-  },
-  content: {},
-}));
