+++ src/routes/Stats.jsx
@@ -1,13 +1,9 @@
 import React from 'react';
-import { DashboardBase } from '../features/dashboard-base';
+import { Layout } from '../components/Layout';
 import { WorldTable } from '../features/world-table';
 import { WorldStatsCards } from '../features/world-stats-cards';
-import Grid from '@material-ui/core/Grid';
-import { GoogleNewsfeed } from '../features/google-newsfeed';
-import { TwitterFeed } from '../features/twitter-feed';
-import { RedditFeed } from '../features/reddit-feed';
 import { makeStyles } from '@material-ui/core/styles';
-
+import { Newsfeed } from '../features/newsfeed';
 
 const useStyles = makeStyles((theme) => ({
     container: {
@@ -24,29 +20,14 @@ function Stats() {
             <WorldTable/>
         </>
     );
-    const renderNewsfeed = () => {
-        return (
-          <div className={classes.container}>
-              <Grid
-                container
-                direction="column"
-                justify="space-between"
-                alignItems="center"
-              >
-              <GoogleNewsfeed/>
-              <TwitterFeed/>
-              <RedditFeed/>
-            </Grid>
-          </div>
-        )
-    }
 
     return (
-        <DashboardBase
-            title={'Worldwide COVID-19 Statistics'}
-            renderContent={renderContent}
-            renderToolbar={renderNewsfeed}
-        />
+        <Layout
+          title={'Worldwide COVID-19 Statistics'}
+          renderToolbar={ () => <Newsfeed/> }
+        >
+          { renderContent() }
+        </Layout>
     )
 }
